[feat]: implement uname syscall, unlock logger on panic and print which process caused crash

+++ Kernel/API/Syscall.hpp
@@ -15,10 +15,10 @@ namespace Syscall
 {
     struct Arguments
     {
-        u64 Index;
-        u64 Args[6];
+        u64       Index;
+        u64       Args[6];
 
-        u64 ReturnValue;
+        uintptr_t ReturnValue;
     };
 
     enum class ID : u64
